テーブル定義書の作成です。（base.pyに全てのテーブルをまとめる）
