+++ back/alembic/versions/20250220-1645_recreate_all_tables.py
@@ -1,8 +1,8 @@
-"""create_tables
+"""recreate_all_tables
 
-Revision ID: bf9ed9e50554
+Revision ID: aa3febf53565
 Revises: 
-Create Date: 2025-02-18 13:44:21.313797
+Create Date: 2025-02-20 16:45:38.248220
 
 """
 from alembic import op
@@ -10,26 +10,19 @@ import sqlalchemy as sa
 
 
 # revision identifiers, used by Alembic.
-revision = 'bf9ed9e50554'
+revision = 'aa3febf53565'
 down_revision = None
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
-    op.create_table('trainings',
-    sa.Column('menu', sa.Text(), nullable=False),
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('created_at', sa.TIMESTAMP(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
-    sa.Column('deleted_at', sa.TIMESTAMP(timezone=True), nullable=True),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('id')
-    )
+    # ### commands auto generated by Alembic - please adjust! ###
     op.create_table('users',
+    sa.Column('id', sa.UUID(), nullable=False),
     sa.Column('email', sa.String(length=255), nullable=False),
     sa.Column('password', sa.String(length=255), nullable=False),
     sa.Column('role', sa.Integer(), nullable=False),
-    sa.Column('id', sa.UUID(), nullable=False),
     sa.Column('created_at', sa.TIMESTAMP(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
     sa.Column('updated_at', sa.TIMESTAMP(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
     sa.Column('deleted_at', sa.TIMESTAMP(timezone=True), nullable=True),
@@ -37,7 +30,24 @@ def upgrade():
     sa.UniqueConstraint('email'),
     sa.UniqueConstraint('id')
     )
+    op.create_table('profiles',
+    sa.Column('id', sa.UUID(), nullable=False),
+    sa.Column('user_id', sa.UUID(), nullable=False),
+    sa.Column('name', sa.String(length=255), nullable=False),
+    sa.Column('team_name', sa.String(length=255), nullable=False),
+    sa.Column('birthday', sa.TIMESTAMP(timezone=True), nullable=False),
+    sa.Column('player_dominant', sa.Enum('RIGHT_RIGHT', 'RIGHT_LEFT', 'LEFT_RIGHT', 'LEFT_LEFT', 'BOTH_RIGHT', 'BOTH_LEFT', 'RIGHT_BOTH', 'LEFT_BOTH', 'BOTH_BOTH', name='dominant_hand_type'), nullable=False),
+    sa.Column('player_position', sa.Enum('PITCHER', 'CATCHER', 'FIRST', 'SECOND', 'THIRD', 'SHORT', 'LEFT', 'CENTER', 'RIGHT', name='player_position_type'), nullable=False),
+    sa.Column('admired_player', sa.String(length=255), nullable=True),
+    sa.Column('introduction', sa.Text(), nullable=True),
+    sa.Column('created_at', sa.TIMESTAMP(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
+    sa.Column('updated_at', sa.TIMESTAMP(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('id')
+    )
     op.create_table('notes',
+    sa.Column('id', sa.UUID(), nullable=False),
     sa.Column('user_id', sa.UUID(), nullable=False),
     sa.Column('theme', sa.String(length=255), nullable=False),
     sa.Column('assignment', sa.Text(), nullable=False),
@@ -47,7 +57,6 @@ def upgrade():
     sa.Column('sleep', sa.DECIMAL(precision=3, scale=1), nullable=False),
     sa.Column('looked_day', sa.Text(), nullable=False),
     sa.Column('practice', sa.Text(), nullable=True),
-    sa.Column('id', sa.UUID(), nullable=False),
     sa.Column('created_at', sa.TIMESTAMP(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
     sa.Column('updated_at', sa.TIMESTAMP(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
     sa.Column('deleted_at', sa.TIMESTAMP(timezone=True), nullable=True),
@@ -55,27 +64,19 @@ def upgrade():
     sa.PrimaryKeyConstraint('id'),
     sa.UniqueConstraint('id')
     )
-    op.create_table('profiles',
-    sa.Column('user_id', sa.UUID(), nullable=False),
-    sa.Column('name', sa.String(length=255), nullable=False),
-    sa.Column('team_name', sa.String(length=255), nullable=False),
-    sa.Column('birthday', sa.TIMESTAMP(timezone=True), nullable=False),
-    sa.Column('player_dominant', sa.Enum('RIGHT_RIGHT', 'RIGHT_LEFT', 'LEFT_RIGHT', 'LEFT_LEFT', 'BOTH_RIGHT', 'BOTH_LEFT', 'RIGHT_BOTH', 'LEFT_BOTH', 'BOTH_BOTH', name='dominant_hand_type'), nullable=False),
-    sa.Column('player_position', sa.Enum('PITCHER', 'CATCHER', 'FIRST', 'SECOND', 'THIRD', 'SHORT', 'LEFT', 'CENTER', 'RIGHT', name='player_position_type'), nullable=False),
-    sa.Column('admired_player', sa.String(length=255), nullable=True),
-    sa.Column('introduction', sa.Text(), nullable=True),
+    op.create_table('trainings',
     sa.Column('id', sa.UUID(), nullable=False),
+    sa.Column('menu', sa.Text(), nullable=False),
     sa.Column('created_at', sa.TIMESTAMP(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
-    sa.Column('updated_at', sa.TIMESTAMP(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
+    sa.Column('deleted_at', sa.TIMESTAMP(timezone=True), nullable=True),
     sa.PrimaryKeyConstraint('id'),
     sa.UniqueConstraint('id')
     )
     op.create_table('comments',
+    sa.Column('id', sa.UUID(), nullable=False),
     sa.Column('user_id', sa.UUID(), nullable=False),
     sa.Column('note_id', sa.UUID(), nullable=False),
     sa.Column('content', sa.Text(), nullable=False),
-    sa.Column('id', sa.UUID(), nullable=False),
     sa.Column('created_at', sa.TIMESTAMP(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
     sa.Column('updated_at', sa.TIMESTAMP(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
     sa.Column('deleted_at', sa.TIMESTAMP(timezone=True), nullable=True),
@@ -85,10 +86,10 @@ def upgrade():
     sa.UniqueConstraint('id')
     )
     op.create_table('training_notes',
+    sa.Column('id', sa.UUID(), nullable=False),
     sa.Column('training_id', sa.UUID(), nullable=False),
     sa.Column('note_id', sa.UUID(), nullable=False),
     sa.Column('count', sa.Integer(), nullable=False),
-    sa.Column('id', sa.UUID(), nullable=False),
     sa.Column('created_at', sa.TIMESTAMP(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
     sa.Column('updated_at', sa.TIMESTAMP(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
     sa.Column('deleted_at', sa.TIMESTAMP(timezone=True), nullable=True),
@@ -104,9 +105,8 @@ def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
     op.drop_table('training_notes')
     op.drop_table('comments')
-    op.drop_table('profiles')
-    op.drop_table('notes')
-    op.drop_table('users')
     op.drop_table('trainings')
+    op.drop_table('notes')
+    op.drop_table('profiles')
+    op.drop_table('users')
     # ### end Alembic commands ###
-
