+++ back/app/models/__init__.py
@@ -2,12 +2,6 @@ import importlib
 import os
 from pathlib import Path
 from app.models.base import Base
-from app.models.user import Users
-from app.models.note import Notes
-from app.models.comment import Comments
-from app.models.profile import Profiles
-from app.models.training import Trainings
-from app.models.training_note import TrainingNotes
